+++ encoding.c
@@ -35,7 +35,7 @@ _bin_search (WCHAR uni, const struct s_uni_trans * tab, WORD end)
 	
 	return (*arr == 0xFFFF ? *(const WCHAR *const*)(arr +1) : arr);
 }
-#define bin_search(uni, tab) _bin_search (uni, tab, numberof(tab) -1)
+#define bin_search(uni, tab) _bin_search (uni, tab, (WORD)numberof(tab) -1)
 
 
 /*==============================================================================
